[DEVOPS-4535] Catch fork up to deleted repo orphans.
The fork for this repo was not up to date with it's upstream. This series of commits will rectify that. The linked commiit in chef-cookbooks was 92aee6313fc47dfd51b45c176b7efe41c767c57d

Tracing that backward to master, the first commit not in this fork is https://github.com/CarnegieLearningWeb/chef-composer/commit/cc64f07695429dd583929417b46350c0b53de779

This commit is cc64f07695429dd583929417b46350c0b53de779